Moved things from cleanup.py into misc.py
